+++ js/util.js
@@ -1,0 +1,36 @@
+// счетчик, создание уникальных id
+const createIdGenerator = () => {
+  let lastGeneratedId = 0;
+
+  return function () {
+    lastGeneratedId += 1;
+    return lastGeneratedId;
+  };
+};
+
+// генератор случайных чисел
+const getRandomInteger = (min, max) => {
+  const lower = Math.ceil(Math.min(min, max));
+  const upper = Math.floor(Math.max(min, max));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+// генератор уникальных случайных чисел
+const getUniqueRandomId = (min, max) => {
+  const previousValues = [];
+
+  return function () {
+    let currentValue = getRandomInteger(min, max);
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomInteger(min, max);
+    }
+    previousValues.push(currentValue);
+    return currentValue;
+  };
+};
+
+//Функция для выбора случайного элемента из массива
+const getRandomArrayElement = (array) => array[getRandomInteger(0, array.length - 1)];
+
+export {createIdGenerator, getRandomInteger, getUniqueRandomId, getRandomArrayElement};
